+++ noether/core/units/__init__.py
@@ -1,6 +1,6 @@
 "Various specialised Units."
 
 from .AffineUnit import AffineUnit
-from .ChainedUnit import ChainedUnit
-from .ComposedUnit import ComposedUnit
+from .LinearUnit import LinearUnit
+from .GeometricUnit import GeometricUnit
 from .PrefixedUnit import PrefixedUnit